# openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Evil Distopia
  description: An API that uses a Gps and medical data
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
paths:
  /users/{user_id}/gps:
    get:
      tags:
        - GPS
      summary: List user's GPS points
      description: Return GPS records for the user.
      operationId: findGps
      parameters:
        - name: user_id
          in: path
          description: ID of User
          required: true
          schema:
            type: string
      responses:
        '200':
          description: gps response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gps'
        '400':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - GPS
      summary: Create GPS point
      description: Create a GPS record for the user; rejects duplicates.
      operationId: addGps
      parameters:
        - name: user_id
          in: path
          description: ID of User
          required: true
          schema:
            type: string
      requestBody:
        description: Gps to add
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGps'
      responses:
        '200':
          description: gps response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gps'
        '400':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: duplicate gps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /gps/{id}:
    get:
      tags:
        - GPS
      summary: Get GPS point
      description: Return a GPS record by id for the user.
      operationId: find Gps by id
      parameters:
        - name: id
          in: path
          description: ID of Gps to fetch
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Gps response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gps'
        '400':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: gps not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
        - GPS
      summary: Update GPS point
      description: Update a GPS record by id for the user.
      operationId: updateGps
      parameters:
        - name: id
          in: path
          description: ID of Gps to update
          required: true
          schema:
            type: string
      requestBody:
        description: Gps to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGps'
      responses:
        '200':
          description: gps response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gps'
        '400':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: gps not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - GPS
      summary: Delete GPS point
      description: Delete a GPS record by id for the user.
      operationId: deleteGps
      parameters:
        - name: id
          in: path
          description: ID of Gps to delete
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Gps deleted
        '400':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: gps not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{user_id}/health_records:
    get:
      tags:
        - Health Records
      summary: List user's medical records
      description: Return medical records for the user.
      operationId: findMedical
      parameters:
        - name: user_id
          in: path
          description: ID of User
          required: true
          schema:
            type: string
      responses:
        '200':
          description: medical response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Medical'
        '400':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Health Records
      summary: Create medical record
      description: Create a medical record for the user; rejects duplicates.
      operationId: addMedical
      parameters:
        - name: user_id
          in: path
          description: ID of User
          required: true
          schema:
            type: string
      requestBody:
        description: Medical to add to the store
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewMedical'
      responses:
        '200':
          description: medical response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medical'
        '400':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: duplicate medical record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /health_records/{id}:
    get:
      tags:
        - Health Records
      summary: Get medical record
      description: Return a medical record by id for the user.
      operationId: find Medical by id
      parameters:
        - name: id
          in: path
          description: ID of Medical to fetch
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Medical response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medical'
        '400':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: medical record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
        - Health Records
      summary: Update medical record
      description: Update a medical record by id for the user.
      operationId: updateMedical
      parameters:
        - name: id
          in: path
          description: ID of Medical to update
          required: true
          schema:
            type: string
      requestBody:
        description: Medical to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewMedical'
      responses:
        '200':
          description: medical response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Medical'
        '400':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: medical record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Health Records
      summary: Delete medical record
      description: Delete a medical record by id for the user.
      operationId: deleteMedical
      parameters:
        - name: id
          in: path
          description: ID of Medical
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Medical deleted
        '400':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: medical record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{user_id}/physical_activities:
    get:
      tags:
        - PHYSICAL ACTIVITY
      summary: List user's physical activities
      description: Return physical activity records for the user.
      operationId: findPhysicalActivity
      parameters:
        - name: user_id
          in: path
          description: ID of physical_activity
          required: true
          schema:
            type: string
      responses:
        '200':
          description: medical response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicalActivities'
        '400':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Gps:
      allOf:
        - $ref: '#/components/schemas/NewGps'
        - type: object
          required:
            - id
            - user_id
          properties:
            id:
              type: string
            user_id:
              type: string
    NewGps:
      type: object
      required:
        - timestamp
        - latitude
        - longitude
      properties:
        timestamp:
          type: string
        latitude:
          type: string
        longitude:
          type: string
    Medical:
      allOf:
        - $ref: '#/components/schemas/NewMedical'
        - type: object
          required:
            - id
            - user_id
          properties:
            id:
              type: string
            user_id:
              type: string
    NewMedical:
      type: object
      required:
        - timestamp
        - heartbeat
      properties:
        timestamp:
          type: string
        heartbeat:
          type: integer
    PhysicalActivity:
      type: object
      required:
        - start_at
        - end_at
        - gps
      properties:
        start_at:
          type: string
        end_at:
          type: string
        itinerary:
          type: array
          items:
            $ref: '#/components/schemas/NewGps'
    PhysicalActivities:
      type: object
      required:
        - user_id
        - physical_activiies
      properties:
        user_id:
          type: string
        physical_activiies:
          type: array
          items:
            $ref: '#/components/schemas/PhysicalActivity'
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string